✨ Feat: add book recommend hook
